djangorestapi/requirements.txt
django
django-admin-tools
djangorestframework
Pillow
psycopg2-binary
django-storages
boto